refactor(compiler): reduce number of reexports of ParseTreeResult
With an upcoming update of the rollup optimizer that is used during bundling, a circular reexport
of ParseTreeResult is discovered.  Instead of expressing reexports of the Symbol within the xml
and html parser files, the ParseTreeResult symbol is explicitly exported in the root
`src/compiler.ts` file.
