+++ packages/compiler/src/ml_parser/xml_parser.ts
@@ -10,8 +10,6 @@ import {TokenizeOptions} from './lexer';
 import {Parser, ParseTreeResult} from './parser';
 import {getXmlTagDefinition} from './xml_tags';
 
-export {ParseTreeResult, TreeError} from './parser';
-
 export class XmlParser extends Parser {
   constructor() {
     super(getXmlTagDefinition);
